# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  product_category.CreateCategoryCommand:
    properties:
      description:
        type: string
      mediaIds:
        items:
          type: integer
        type: array
      name:
        type: string
      order:
        type: integer
      parentCategoryId:
        type: integer
      seoTags:
        items:
          type: string
        type: array
      siteId:
        type: integer
      slug:
        type: string
    required:
    - order
    - siteId
    - slug
    type: object
  product_category.DeleteCategoryCommand:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  product_category.UpdateCategoryCommand:
    properties:
      description:
        type: string
      id:
        type: integer
      mediaIds:
        items:
          type: integer
        type: array
      name:
        type: string
      order:
        type: integer
      parentCategoryId:
        type: integer
      seoTags:
        items:
          type: string
        type: array
      siteId:
        type: integer
      slug:
        type: string
    required:
    - id
    - order
    - siteId
    type: object
...
  utils.Result:
    properties:
      data: {}
      message:
        type: string
      statusCode:
        type: integer
      success:
        type: boolean
      systemMessage:
        type: string
    type: object
...
host: localhost:8585
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: This is a sample server.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: My API
  version: "1.0"
paths:
...
  /product-category:
    delete:
      consumes:
      - application/json
      description: Deletes an existing product category by its ID
      parameters:
      - description: Category ID to delete
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product_category.DeleteCategoryCommand'
      produces:
      - application/json
      responses:
        "200":
          description: Deleted category confirmation
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: Delete a product category
      tags:
      - product-category
    get:
      consumes:
      - application/json
      description: Retrieves a specific product category by its ID
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category details
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: Get product category by ID
      tags:
      - product-category
    post:
      consumes:
      - application/json
      description: Creates a new category for products with the provided information
      parameters:
      - description: Category information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product_category.CreateCategoryCommand'
      produces:
      - application/json
      responses:
        "201":
          description: Created category
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: Create a new product category
      tags:
      - product-category
    put:
      consumes:
      - application/json
      description: Updates an existing product category with the provided information
      parameters:
      - description: Updated category information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product_category.UpdateCategoryCommand'
      produces:
      - application/json
      responses:
        "200":
          description: Updated category
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: Update a product category
      tags:
      - product-category
  /product-category/admin/all:
    get:
      consumes:
      - application/json
      description: Admin endpoint to retrieve all product categories with additional
        information
      parameters:
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      - in: query
        name: search
        type: string
      - enum:
        - title
        - description
        in: query
        name: searchBy
        type: string
      - enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - enum:
        - title
        - description
        in: query
        name: sortBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of all categories
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: 'Admin: Get all product categories'
      tags:
      - product-category
  /product-category/all:
    get:
      consumes:
      - application/json
      description: Retrieves all product categories with optional filtering
      parameters:
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      - in: query
        name: search
        type: string
      - enum:
        - title
        - description
        in: query
        name: searchBy
        type: string
      - in: query
        name: siteId
        required: true
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - enum:
        - title
        - description
        in: query
        name: sortBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of categories
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: Get all product categories
      tags:
      - product-category
...
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"